+++ src/createMappedRegion/__test__/public.test.ts
@@ -86,29 +86,3 @@ describe('load', () => {
         expect(result).toBe(2);
     });
 });
-
-describe('getReducedValue', () => {
-    test('getProps', () => {
-        const reducer = (state: any, params: string) => {
-            const {value, pendingMutex} = state[params] ?? {};
-            return {loading: pendingMutex !== 0, value};
-        };
-        expect(region.getReducedValue('uniq1', reducer)).toEqual({value: undefined, loading: true});
-
-        region.set('uniq1', 1);
-        expect(region.getReducedValue('uniq1', reducer)).toEqual({value: 1, loading: false});
-    });
-
-    test('getListValue', () => {
-        const reducer = (state: any, params: string[]) => {
-            return params.map(key => (state[key] ?? {}).value);
-        };
-        expect(region.getReducedValue(['a', 'b'], reducer)).toEqual([undefined, undefined]);
-
-        region.set('a', 1);
-        region.set('b', 2);
-        region.set('c', 3);
-
-        expect(region.getReducedValue(['a', 'b'], reducer)).toEqual([1, 2]);
-    });
-});